requirements.txt
python-telegram-bot==20.3
pandas==2.0.3
lxml==4.9.3
requests==2.31.0
python-dotenv==1.0.0
openpyxl==3.1.2